Sat, Apr 9, 2022, 12:55 PM +00:00

+++ caesar/caesar.c
@@ -55,7 +55,7 @@ int valid_input(string key)
                 return 1;
             }
     }
-    return (int) key;
+    return key;
 }
